Remove autologin configuration if exists

+++ DEBIAN13/tools/adduser.sh
@@ -17,8 +17,13 @@ if [[ "$SUDO" == "yes" ]]; then
  sudo adduser $USERNAME sudo
 fi
 
+if [[ -f /etc/sddm.conf.d/autologin.conf ]]; then
 sudo rm -f /etc/sddm.conf.d/autologin.conf
+fi
 
+if [[ -f /etc/lightdm/lightdm.conf.d/49-autologin.conf ]]; then
+sudo rm -f /etc/lightdm/lightdm.conf.d/49-autologin.conf
+fi
 
 
 # === Autologin fragen ===
